Add files via upload
Objective: create a multiset to hold pointers to person object. Define the multiset w/ the comparePerson function object, so it can be sorted automatically by names of person.  Several of the persons should have the same name, to verify that the multiset stores multiple object with the same key. And allow the user to search a person object by last name and the first name.